pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "exam-helper"
version = "0.1.0"
description = "Local-first exam authoring helper for physics courses."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "jinja2>=3.1.0",
  "python-multipart>=0.0.20",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "python-dotenv>=1.1.0",
  "pyyaml>=6.0.2",
  "sympy>=1.14.0",
  "pint>=0.24.0",
  "openai>=1.99.0",
  "python-docx>=1.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.28.0",
]

[project.scripts]
exam-helper = "exam_helper.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/exam_helper"]

[tool.hatch.build.targets.wheel.force-include]
"src/exam_helper/templates" = "exam_helper/templates"
"src/exam_helper/prompt_templates.yaml" = "exam_helper/prompt_templates.yaml"

[tool.hatch.build.targets.sdist.force-include]
"src/exam_helper/templates" = "src/exam_helper/templates"
"src/exam_helper/prompt_templates.yaml" = "src/exam_helper/prompt_templates.yaml"
